gralloc.h: build fix
Fixes building by #ifndef guarding a possible double declaration of
GRALLOC_MODULE_PERFORM_CREATE_HANDLE_FROM_BUFFER in our local gralloc.h.

The file is included a few times and in the case when it is called from
hardware/qcom/display/libgralloc/gralloc_priv.h,
GRALLOC_MODULE_PERFORM_CREATE_HANDLE_FROM_BUFFER is already defined
there, leading to a double declaration.  However, this definition is
needed by other parts in the code so we can not simply remove it, but
instead guard the definition with #ifndef that skips it only when the
file is included from gralloc_priv.h.